Move non analysis code to StatementGenerator class
