+++ core/src/main/java/novoda/lib/sqliteprovider/util/analyzer/DatabaseAnalyzer.java
@@ -1,7 +1,6 @@
 
 package novoda.lib.sqliteprovider.util.analyzer;
 
-import android.content.ContentValues;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 
@@ -10,11 +9,9 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
 
 import novoda.lib.sqliteprovider.sqlite.IDatabaseMetaInfo.SQLiteType;
 import novoda.lib.sqliteprovider.util.Constraint;
-import novoda.rest.database.SQLiteTableCreator;
 
 public class DatabaseAnalyzer {
 
@@ -125,14 +122,14 @@ public class DatabaseAnalyzer {
         return version;
     }
 
-    public List<Constraint> getUniqueConstraints(SQLiteDatabase db, String table) {
+    public List<Constraint> getUniqueConstraints(String table) {
         List<Constraint> constraints = new ArrayList<Constraint>();
-        final Cursor indexCursor = db.rawQuery(String.format(PRAGMA_INDEX_LIST, table), null);
+        final Cursor indexCursor = executeQuery(String.format(PRAGMA_INDEX_LIST, table));
         while (indexCursor.moveToNext()) {
             int isUnique = indexCursor.getInt(2);
             if (isUnique == 1) {
                 String indexName = indexCursor.getString(1);
-                final Cursor columnCursor = db.rawQuery(String.format(PRAGMA_INDEX_INFO, indexName), null);
+                final Cursor columnCursor = executeQuery(String.format(PRAGMA_INDEX_INFO, indexName));
                 List<String> columns = new ArrayList<>(columnCursor.getCount());
                 while (columnCursor.moveToNext()) {
                     String columnName = columnCursor.getString(2);
@@ -142,69 +139,9 @@ public class DatabaseAnalyzer {
                 constraints.add(new Constraint(columns));
             }
         }
+
         indexCursor.close();
+
         return constraints;
-    }
-
-    public String contentValuestoTableCreate(ContentValues values, String table) {
-        StringBuilder string = new StringBuilder("CREATE TABLE ")
-                .append(table)
-                .append(" (");
-
-        for (Entry<String, Object> entry : values.valueSet()) {
-            string.append(entry.getKey())
-                    .append(" TEXT, ");
-        }
-
-        return string.delete(string.length() - 2, string.length())
-                .append(");")
-                .toString();
-    }
-
-    public String getCreateStatement(SQLiteTableCreator creator) {
-
-        String primaryKeyColumnName = creator.getPrimaryKey();
-        novoda.rest.database.SQLiteType primaryKeyColumnType;
-        boolean shouldAutoincrement;
-        if (primaryKeyColumnName == null) {
-            primaryKeyColumnName = "_id";
-            primaryKeyColumnType = novoda.rest.database.SQLiteType.INTEGER;
-            shouldAutoincrement = true;
-        } else {
-            primaryKeyColumnType = creator.getType(primaryKeyColumnName);
-            shouldAutoincrement = creator.shouldPKAutoIncrement();
-        }
-
-
-        StringBuilder sql = new StringBuilder().append("CREATE TABLE IF NOT EXISTS ")
-                .append("\"").append(creator.getTableName()).append("\"")
-                .append(" (").append(primaryKeyColumnName).append(" ").append(primaryKeyColumnType.name())
-                .append(" PRIMARY KEY ");
-
-        if (shouldAutoincrement) {
-            sql.append("AUTOINCREMENT ");
-        }
-
-        for (String columnName : creator.getTableFields()) {
-            if (columnName.equals(primaryKeyColumnName)) {
-                continue;
-            }
-            sql.append(", ").append(columnName).append(" ").append(creator.getType(columnName).name());
-
-            if (!creator.isNullAllowed(columnName)) {
-                sql.append(" NOT NULL");
-            }
-
-            if (creator.isUnique(columnName)) {
-                sql.append(" UNIQUE");
-            }
-
-            if (creator.onConflict(columnName) != null && creator.isUnique(columnName)) {
-                sql.append(" ON CONFLICT ").append(creator.onConflict(columnName));
-            }
-        }
-
-        sql.append(");");
-        return sql.toString();
     }
 }
